requirements.txt
biopython==1.78
gffpandas==0.3.9
numpy==1.21.0
pandas==1.3.0
matplotlib==3.4.3
seaborn==0.11.2
pyarrow==5.0.0  # Add this line if you are using pyarrow for Parquet files